'tab' is the most important key in python!

+++ python/binary_search.py
@@ -2,7 +2,7 @@ def binary_search(list,item):
     low = 0
     high = len(list) - 1
     while low <= high:
-        mid = (low + high)/2
+        mid = int((low + high)/2)
         guess = list[mid]
         if guess == item:
             return mid
@@ -10,7 +10,7 @@ def binary_search(list,item):
             high = mid - 1
         else:
             low = mid + 1
-        return None
+    return None
 
 ##############################
